Add CNAME file (#17)
Create CNAME

CNAME is needed when you publish from a (gh-pages) branch with custom domain. In that case, the GitHub settings only creates a commit on the specified branch. In our case, gh-pages is build from main, but main does not have CNAME so the GitHub-created one was removed on every update. This adds the CNAME file added by GitHub to gh-pages to main.