CI: remove _freeze to build the entire page (#42)
To simplify contribution workflow, I removed the _freeze folder so the CI will need to render the entire page. Otherwise it's difficult for contributors always add the changes in the freeze folder, especially when fixing typos etc.

We don't evaluate cells in any of the tutorials. If we have a specific one that will need to be precomputed, we will use the freezed content only for that specific tutorials.

I also added automatic notebook output for some tutorials. It shows up in TOC